Backend/requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dotenv==0.9.9
Flask==3.1.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.6.0
numpy==2.4.2
pandas==3.0.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.46
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
Werkzeug==3.1.6